// src/gamma-engine/slots/model/event/SlotMachineEvent.ts
export enum SlotMachineEvent {
    STATE_CHANGED = 'onSlotMachineStateChanged',
    BET_VALUE_CHANGED = 'onSlotMachineBetValueChanged',
    GAME_SPEED_LEVEL_CHANGED = 'onSlotMachineGameSpeedLevelChanged',
    NEW_LINE_SHOW = 'onNewLineShow',
    BALANCE_CHANGED = 'onSlotMachineBalanceChanged',
    KEY_PRESSED = 'onKeyPressed',
}